Move `TimeFormatter` to its own file in production
